requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aniso8601==6.0.0          # via graphene
appdirs==1.4.3            # via black
argon2-cffi==19.1.0
aspy.yaml==1.3.0          # via pre-commit
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via black, pytest
black==19.3b0
cffi==1.12.3              # via argon2-cffi
cfgv==2.0.0               # via pre-commit
click==7.0                # via black, pip-tools
django==2.2.2
djangorestframework-simplejwt==4.3.0
djangorestframework==3.9.4  # via djangorestframework-simplejwt
entrypoints==0.3          # via flake8
flake8==3.7.7
graphene-django==2.3.0
graphene==2.1.6           # via graphene-django
graphql-core==2.2         # via graphene, graphene-django, graphql-relay
graphql-relay==0.4.5      # via graphene
identify==1.4.5           # via pre-commit
importlib-metadata==0.18  # via pluggy, pre-commit, pytest
mccabe==0.6.1             # via flake8
more-itertools==7.0.0     # via pytest
nodeenv==1.3.3            # via pre-commit
packaging==19.0           # via pytest
pip-tools==3.8.0
pluggy==0.12.0            # via pytest
pre-commit==1.17.0
promise==2.2.1            # via graphene-django, graphql-core, graphql-relay
psycopg2-binary==2.8.3
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via flake8
pycparser==2.19           # via cffi
pyflakes==2.1.1           # via flake8
pyjwt==1.7.1              # via djangorestframework-simplejwt
pyparsing==2.4.0          # via packaging
pytest-django==3.5.0
pytest==4.6.3             # via pytest-django
python-dotenv==0.10.3
pytz==2019.1              # via django
pyyaml==5.1.1             # via aspy.yaml, pre-commit
rx==1.6.1                 # via graphql-core
singledispatch==3.4.0.3   # via graphene-django
six==1.12.0               # via argon2-cffi, cfgv, graphene, graphene-django, graphql-core, graphql-relay, packaging, pip-tools, pre-commit, promise, pytest, singledispatch
sqlparse==0.3.0           # via django
toml==0.10.0              # via black, pre-commit
virtualenv==16.6.0        # via pre-commit
wcwidth==0.1.7            # via pytest
zipp==0.5.1               # via importlib-metadata
